# backend/mja-api/src/main/resources/META-INF/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: mja-api
  description: API für mein Internet-Aufgabenarchiv
  contact:
    name: Heike Winkelvoß
    url: https://mathe-jung-alt.de
    email: info@egladil.de
  license:
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.4.0
servers:
- url: http://localhost:9210
  description: Auto generated value
- url: http://0.0.0.0:9210
  description: Auto generated value
...
paths:
...
  /quellen/admin/v1:
    get:
      tags:
      - Quellen
      summary: "Gibt die Quelle zurück, die zu der eingeloggten Person gehört"
      operationId: findOwnQuelle
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quelle'
        "404":
          description: Gibt es nicht
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      deprecated: true
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /quellen/admin/v2:
    get:
      tags:
      - Quellen
      summary: "Gibt die Quelle zurück, die zu der eingeloggten Person gehört"
      operationId: getQuelleEingeloggterAdmin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quelle'
        "404":
          description: Gibt es nicht
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /quellen/v1:
    get:
      tags:
      - Quellen
      summary: "Gibt alle Quellen zurück, die auf die gegebene Suchanfrage passen."
      operationId: findQuellenV1
      parameters:
      - name: suchstring
        in: query
        description: Freitext zum suchen. Es wird mit like nach diesem Text gesucht
        schema:
          pattern: "^[\\w äöüß \\+ \\- \\. \\,]{1,30}$"
          type: string
      - name: deskriptoren
        in: query
        description: kommaseparierte Liste von Namen von Deskriptoren
        schema:
          pattern: "^[\\d\\,]{0,200}$"
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quelle'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      deprecated: true
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /quellen/v1/{quelleId}:
    get:
      tags:
      - Quellen
      summary: Gibt die Quelle mit der gegebenen ID zurück
      operationId: findQuelleById
      parameters:
      - name: quelleId
        in: path
        description: technische ID der Quelle
        required: true
        schema:
          pattern: "^[a-fA-F\\d\\-]{1,36}$"
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quelle'
        "404":
          description: Gibt es nicht
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /quellen/v2:
    get:
      tags:
      - Quellen
      summary: "Gibt alle Quellen zurück, die auf die gegebene Suchanfrage passen."
      operationId: findQuellen
      parameters:
      - name: suchstring
        in: query
        description: Freitext zum suchen. Es wird mit like nach diesem Text gesucht
        schema:
          pattern: "^[\\w äöüß \\+ \\- \\. \\,]{1,30}$"
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quelle'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
...
components:
  schemas:
...
    MessagePayload:
      description: eine kategorisierte Meldung
      type: object
      properties:
        level:
          description: "Level der Message: INFO | WARN | ERROR"
          type: string
        message:
          description: die message
          type: string
...
    Quelle:
      description: Quelle für ein Rätsel
      type: object
      properties:
        id:
          description: "technische ID, 'neu' für neue Rätsel"
          type: string
        quellenart:
          description: "Art der Quelle: Mensch, Buch, Zeitschrift"
          type: string
          allOf:
          - $ref: '#/components/schemas/Quellenart'
        sortNumber:
          format: int64
          description: Zahl zum Sortieren
          type: integer
        name:
          description: menschenlesbarer Anzeigetext für eine Quellenangabe
          type: string
        mediumUuid:
          description: Referenz auf ein Buch oder eine Zeitschrift
          type: string
        deskriptoren:
          description: "Deskriptoren, für die Quelle"
          type: array
          items:
            description: Deskriptor-Entity mit allen Attributen.
            type: object
            properties:
              id:
                format: int64
                description: technische ID
                type: integer
              name:
                description: Name zum Anzeigen
                type: string
              admin:
                description: "Flag, ob dieser Deskriptor nur in der Admin-Anwendung\
                  \ zur Verfügung steht"
                type: boolean
              kontext:
                description: Kontext des Deskriptors
                type: string
                deprecated: true
...
    Quellenart:
      enum:
      - BUCH
      - PERSON
      - ZEITSCHRIFT
      type: string
...